API

List items

QUBOLib.list_collections — Function

list_collections()

List the codes of the registered collections.

QUBOLib.list_instances — Function

list_instances(collection::AbstractString)

Lists the codes for all instances in a collection.

Load instances

QUBOLib.load_instance — Function

load_instance(collection::AbstractString, instance::AbstractString)

Loads a specific instance for a given collection as a QUBOTools model.