API

Actions

Path Routing

QUBOLib.library_path — Function

library_path()::AbstractString

Returns the absolute path to the QUBOLib artifact.

QUBOLib.database_path — Function

database_path(path::AbstractString=library_path())::AbstractString

Returns the absolute path to the database file, given a reference path.

QUBOLib.archive_path — Function

archive_path(path::AbstractString=library_path())::AbstractString

Returns the absolute path to the archive file, given a reference path.

QUBOLib.root_path — Function

root_path()::AbstractString

Returns the absolute path to the project's root folder.

Info

The dist_path, build_path, and cache_path functions are more often used when building the library, therefore they will point to the the project's root path by default, by referencing this function.

QUBOLib.dist_path — Function

dist_path(path::AbstractString=root_path())::AbstractString

Returns the absolute path to the distribution folder, given a reference path. The path is created if it does not exist.

QUBOLib.build_path — Function

build_path(path::AbstractString=root_path())::AbstractString

Returns the absolute path to the build folder, given a reference path. The path is created if it does not exist.

QUBOLib.cache_path — Function

cache_path(path::AbstractString=root_path())::AbstractString

Returns the absolute path to the cache folder, given a reference path. The path is created if it does not exist.

Library Index

QUBOLib.LibraryIndex — Type

LibraryIndex

The QUBOLib index is composed of two pieces: a SQLite database and an HDF5 archive.

QUBOLib.database — Function

database(index::LibraryIndex)::SQLite.DB

Returns a pointer that grants direct access to the SQLite database of the library index.

QUBOLib.archive — Function

archive(index::LibraryIndex)::HDF5.File

Returns a pointer that grants direct access to the HDF5 archive of the library index.

Data Access

QUBOLib.access — Function

access(
    callback;
    path::Union{AbstractString,Nothing} = nothing,
    create::Bool = false
)::LibraryIndex

Loads the index for an instance library.

If path is not provided, the latest QUBOLib artifact will be used.

Example

using QUBOLib

QUBOLib.access() do index
    print(index) # Show some information about the index
end

QUBOLib.load_collection — Function

load_collection(index::LibraryIndex, code::Symbol)

QUBOLib.load_instance — Function

load_instance(index::LibraryIndex, instance::Integer)

QUBOLib.load_solution — Function

load_solution(index::LibraryIndex, solution::Integer)
load_solution(index::LibraryIndex, instance::Integer, solution::Integer)

Data Management

QUBOLib.add_collection! — Function

add_collection!(index::LibraryIndex, code::Symbol, data::Dict{String,Any})

Creates a new collection in the library index.

QUBOLib.add_instance! — Function

add_instance!(index::LibraryIndex, coll::Symbol, model::QUBOTools.Model{Int,Float64,Int})

Adds a new instance to the library index.

QUBOLib.add_solution! — Function

add_solution!(index::LibraryIndex, instance::Integer, solution::SampleSet{Float64,Int})

Registers a new solution for a given instance.

The solution argument is a QUBOTools.SampleSet, which is a collection of samples and their respective energies.

QUBOLib.add_solver! — Function

add_solver!(index::LibraryIndex, code::Symbol, data::Dict{String,Any})

Registers a new solver in the library index.

QUBOLib.remove_collection! — Function

remove_collection!(index::LibraryIndex, code::Symbol)

Removes a collection and its contents from the library index.

QUBOLib.remove_instance! — Function

remove_instance!(index::LibraryIndex, coll::Symbol, instance::Integer)

Removes an instance from the library index.

QUBOLib.remove_solution! — Function

remove_solution!(index::LibraryIndex, instance::Integer, solution::Integer)

Removes a solution from the library index.

QUBOLib.remove_solver! — Function

remove_solver!(index::LibraryIndex, code::Symbol)

Removes a solver from the library index.

Instance Synthesis

QUBOLib.Synthesis.AbstractProblem — Type

AbstractProblem{T}

QUBOLib.Synthesis.generate — Function

generate(problem::AbstractProblem{T}) where {T}
generate(rng, problem::AbstractProblem{T}) where {T}

Generates a QUBO problem and returns it as a QUBOTools.Model.

Problem Types

QUBOLib.Synthesis.NAE3SAT — Type

NAE3SAT{T}(m::Integer, n::Integer)

Not-all-equal 3-SAT on $m$ clauses and $n$ variables.

QUBOLib.Synthesis.XORSAT — Type

XORSAT{T}(n::Integer, r::Integer = 3, k::Integer = 3)

$r$-regular $k$-XORSAT on $n$ variables.

QUBOLib.Synthesis.Wishart — Type

Wishart{T}(n::Integer, m::Integer)

Represents the Wishart model on $n$ variables whose $\mathbf{W}$ matrix has $m$ columns.

When true, the discretize keyword limits the entries of the $\mathbf{R}$ matrix to $\pm 1$. The precision, on the other hand, is the amount of digits to round each entry $R_{i,j}$ after sampling from a normal distribution $\mathcal{N}(0, 1)$.

QUBOLib.Synthesis.SherringtonKirkpatrick — Type

SherringtonKirkpatrick{T}(n::Integer, μ::T, σ::T)

Generates a Sherrington-Kirkpatrick model in $n$ variables. Coefficients are normally distributed with mean $\mu$ and variance $\sigma$.